Restruturação e complementação do código
- Troca efetiva de variáveis 'int' para 'quint8'
- Ajuste  no comunicação do AMF para permitir troca de passes
- Adaptação na Skill AroundTheBall e no Playbook DoNothing

+++ entity/player/behaviour/basics/behaviour_ballreceptor.h
@@ -32,7 +32,7 @@ private:
     int _state;
 
     //Parameters
-    int _id;
+    quint8 _id;
 
     //Skills
     //Skill_RotateTo *_sk_rotateto;
@@ -41,7 +41,7 @@ public:
     Behaviour_BallReceptor();
     QString name();
 
-    void setPlayerId (int id) { _id = id; }
+    void setPlayerId (quint8 id) { _id = id; }
 };
 
 #endif // BEHAVIOUR_BALLRECEPTOR_H
