+++ entity/player/behaviour/basics/behaviour_cover.h
@@ -34,14 +34,14 @@ private:
     Skill_GoToLookTo *_skill_GoToLookTo;
 
     // Parameters
-    int _id;
+    quint8 _id;
     float _a, _b, _c;
 
 public:
     Behaviour_Cover();
     QString name();
 
-    void setIdOfPoss (int idWithPoss) { _id = idWithPoss; }
+    void setIdOfPoss (quint8 idWithPoss) { _id = idWithPoss; }
     void setACoeficient (float a) { _a = a; }
     void setBCoeficient (float b) { _b = b; }
     void setCCoeficient (float c) { _c = c; }
