+++ entity/player/role/basics/role_def_midfielder.cpp
@@ -47,7 +47,7 @@ void Role_Def_Midfielder::run(){
     //printf("ID COM POSSE: %i\n", idWithPoss);
 
     if (ourPoss == false) {
-        int idWithPoss = playerWithPoss(ourPoss);
+        quint8 idWithPoss = playerWithPoss(ourPoss);
         if (idWithPoss == BALLPOSS_NONE) {
             if (_isPassComing == true) {
                 _bh_brp->setPlayerId(player()->playerId());
@@ -63,10 +63,10 @@ void Role_Def_Midfielder::run(){
             }
         }
     } else {
-        int idWithPoss = playerWithPoss(ourPoss);
+        quint8 idWithPoss = playerWithPoss(ourPoss);
         if (idWithPoss == player()->playerId()) {
             _bh_psg->setPlayerId(idWithPoss);
-            int passId = _bh_psg->getPassId();
+            quint8 passId = _bh_psg->getPassId();
             emit sendPassId(passId);
             setBehaviour(BHV_PASSING);
             previousPoss = true;
@@ -96,7 +96,7 @@ bool Role_Def_Midfielder::ourTeamPossession() {
     return false;
 }
 
-int Role_Def_Midfielder::playerWithPoss(bool ourPoss) {
+quint8 Role_Def_Midfielder::playerWithPoss(bool ourPoss) {
     if (ourPoss == true) {
         for (quint8 i = 0; i < MRCConstants::_qtPlayers; i++) {
             if (PlayerBus::ourPlayerAvailable(i)) {
@@ -119,7 +119,7 @@ int Role_Def_Midfielder::playerWithPoss(bool ourPoss) {
     return BALLPOSS_NONE;
 }
 
-void Role_Def_Midfielder::receivePassId(int passId) {
+void Role_Def_Midfielder::receivePassId(quint8 passId) {
     if (passId == player()->playerId()) {
         _isPassComing = true;
         //std::cout << "[DMF] Aqui!\n";
