+++ entity/player/role/basics/role_defender.h
@@ -53,13 +53,13 @@ private:
 
     // Auxiliary functions
     bool ourTeamPossession();
-    int playerWithPoss(bool ourPoss);
+    quint8 playerWithPoss(bool ourPoss);
 
 signals:
-    void sendPassId(int passId);
+    void sendPassId(quint8 passId);
 
 public slots:
-    void receivePassId(int passId);
+    void receivePassId(quint8 passId);
 
 public:
     Role_Defender();
